Phase 2: Gating python specific code and move python bytecode logic into separate mod
