+++ crates/qasm2/Cargo.toml
@@ -12,10 +12,15 @@ doctest = false
 [lints]
 workspace = true
 
+[features]
+default = ["py"]
+py = ["dep:pyo3"]
+
 [dependencies]
 num-bigint.workspace = true
 hashbrown.workspace = true
 
 [dependencies.pyo3]
 workspace = true
+optional = true
 features = ["py-clone", "num-bigint"]
