+++ crates/qasm2/src/parse.rs
@@ -17,6 +17,7 @@
 
 use hashbrown::{HashMap, HashSet};
 use num_bigint::BigUint;
+#[cfg(feature = "py")]
 use pyo3::prelude::*;
 
 use crate::bytecode::InternalBytecode;
@@ -25,7 +26,7 @@ use crate::error::{
 };
 use crate::expr::{Expr, ExprParser};
 use crate::lex::{Token, TokenContext, TokenStream, TokenType, Version};
-use crate::{ClassicalCallableExt, CustomClassical, CustomInstruction};
+use crate::{Attachment, ClassicalCallableExt, CustomClassical, CustomInstruction};
 
 /// The number of gates that are built in to the OpenQASM 2 language.  This is U and CX.
 const N_BUILTIN_GATES: usize = 2;
@@ -64,7 +65,8 @@ const BUILTIN_CLASSICAL: [&str; 6] = ["cos", "exp", "ln", "sin", "sqrt", "tan"];
 /// the second is whether to also define addition to make offsetting the newtype easier.
 macro_rules! newtype_id {
     ($id:ident, false) => {
-        #[derive(Clone, Copy, Debug, PartialEq, Eq, Hash, IntoPyObject, IntoPyObjectRef)]
+        #[derive(Clone, Copy, Debug, PartialEq, Eq, Hash)]
+        #[cfg_attr(feature = "py", derive(IntoPyObject, IntoPyObjectRef))]
         pub struct $id(usize);
 
         impl $id {
@@ -689,6 +691,7 @@ impl State {
     fn parse_gate_definition(
         &mut self,
         bc: &mut Vec<Option<InternalBytecode>>,
+        attachment: Attachment<'_>,
     ) -> Result<usize, ParseError> {
         let gate_token = self.expect_known(TokenType::Gate);
         let name_token = self.expect(TokenType::Id, "an identifier", &gate_token)?;
@@ -783,7 +786,9 @@ impl State {
         let mut statements = 0usize;
         loop {
             match self.peek_token()?.map(|tok| tok.ttype) {
-                Some(TokenType::Id) => statements += self.parse_gate_application(bc, None, true)?,
+                Some(TokenType::Id) => {
+                    statements += self.parse_gate_application(bc, None, true, attachment)?
+                }
                 Some(TokenType::Barrier) => {
                     statements += self.parse_barrier(bc, Some(num_qubits))?
                 }
@@ -893,6 +898,7 @@ impl State {
         bc: &mut Vec<Option<InternalBytecode>>,
         condition: Option<Condition>,
         in_gate: bool,
+        attachment: Attachment<'_>,
     ) -> Result<usize, ParseError> {
         let name_token = self.expect_known(TokenType::Id);
         let name = name_token.id(&self.context);
@@ -920,7 +926,8 @@ impl State {
                 Err(ParseError::new(message_generic(Some(&pos), &message)))
             }
         }?;
-        let parameters = self.expect_gate_parameters(&name_token, num_params, in_gate)?;
+        let parameters =
+            self.expect_gate_parameters(&name_token, num_params, in_gate, attachment)?;
         let mut qargs = Vec::<Operand<QubitId>>::with_capacity(num_qubits);
         let mut comma = None;
         if in_gate {
@@ -977,13 +984,18 @@ impl State {
     }
 
     /// Parse an expected expression at this position.
-    fn expect_expression(&mut self, cause: &Token) -> Result<Expr, ParseError> {
+    fn expect_expression(
+        &mut self,
+        cause: &Token,
+        attachment: Attachment<'_>,
+    ) -> Result<Expr, ParseError> {
         ExprParser {
             tokens: &mut self.tokens,
             context: &mut self.context,
             gate_symbols: &self.gate_symbols,
             global_symbols: &self.symbols,
             strict: self.strict,
+            attachment,
         }
         .parse_expression(cause)
     }
@@ -994,6 +1006,7 @@ impl State {
         name_token: &Token,
         num_params: usize,
         in_gate: bool,
+        attachment: Attachment<'_>,
     ) -> Result<GateParameters, ParseError> {
         let lparen_token = match self.accept(TokenType::LParen)? {
             Some(lparen_token) => lparen_token,
@@ -1013,7 +1026,7 @@ impl State {
         let parameters = if in_gate {
             let mut parameters = Vec::<Expr>::with_capacity(num_params);
             while !self.next_is(TokenType::RParen)? {
-                parameters.push(self.expect_expression(&lparen_token)?);
+                parameters.push(self.expect_expression(&lparen_token, attachment)?);
                 seen_params += 1;
                 comma = self.accept(TokenType::Comma)?;
                 if comma.is_none() {
@@ -1025,7 +1038,7 @@ impl State {
         } else {
             let mut parameters = Vec::<f64>::with_capacity(num_params);
             while !self.next_is(TokenType::RParen)? {
-                match self.expect_expression(&lparen_token)? {
+                match self.expect_expression(&lparen_token, attachment)? {
                     Expr::Constant(value) => parameters.push(value),
                     _ => {
                         return Err(ParseError::new(message_generic(
@@ -1252,6 +1265,7 @@ impl State {
     fn parse_conditional(
         &mut self,
         bc: &mut Vec<Option<InternalBytecode>>,
+        attachment: Attachment<'_>,
     ) -> Result<usize, ParseError> {
         let if_token = self.expect_known(TokenType::If);
         let lparen_token = self.expect(TokenType::LParen, "'('", &if_token)?;
@@ -1287,7 +1301,7 @@ impl State {
         }?;
         let condition = Some(Condition { creg, value });
         match self.peek_token()?.map(|tok| tok.ttype) {
-            Some(TokenType::Id) => self.parse_gate_application(bc, condition, false),
+            Some(TokenType::Id) => self.parse_gate_application(bc, condition, false, attachment),
             Some(TokenType::Measure) => self.parse_measure(bc, condition),
             Some(TokenType::Reset) => self.parse_reset(bc, condition),
             Some(_) => {
@@ -1725,6 +1739,7 @@ impl State {
     pub fn parse_next(
         &mut self,
         bc: &mut Vec<Option<InternalBytecode>>,
+        attachment: Attachment<'_>,
     ) -> Result<Option<usize>, ParseError> {
         if self.strict && self.allow_version {
             match self.peek_token()?.map(|tok| tok.ttype) {
@@ -1751,16 +1766,16 @@ impl State {
         self.allow_version = false;
         while let Some(ttype) = self.peek_token()?.map(|tok| tok.ttype) {
             let emitted = match ttype {
-                TokenType::Id => self.parse_gate_application(bc, None, false)?,
+                TokenType::Id => self.parse_gate_application(bc, None, false, attachment)?,
                 TokenType::Creg => self.parse_creg(bc)?,
                 TokenType::Qreg => self.parse_qreg(bc)?,
                 TokenType::Include => self.parse_include(bc)?,
                 TokenType::Measure => self.parse_measure(bc, None)?,
                 TokenType::Reset => self.parse_reset(bc, None)?,
                 TokenType::Barrier => self.parse_barrier(bc, None)?,
-                TokenType::If => self.parse_conditional(bc)?,
+                TokenType::If => self.parse_conditional(bc, attachment)?,
                 TokenType::Opaque => self.parse_opaque_definition(bc)?,
-                TokenType::Gate => self.parse_gate_definition(bc)?,
+                TokenType::Gate => self.parse_gate_definition(bc, attachment)?,
                 TokenType::OpenQASM => {
                     if allow_version {
                         self.parse_version()?
